Adjust scanning functionality, add stock browser with QR codes

+++ backend/src/main/java/com/farmstock/model/StockAdjustmentRequest.java
@@ -2,10 +2,12 @@ package com.farmstock.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class StockAdjustmentRequest {
-    private Double quantity;
-    private String movementType;  // IN, OUT, ADJUSTMENT
+    private BigDecimal quantity;
+    private String movementType;
     private String reason;
     private Long userId;
 }
